# docs/swagger.yaml
definitions:
...
  models.Login:
    properties:
      code:
        type: string
      password:
        type: string
      username:
        type: string
      uuid:
        type: string
    required:
    - code
    - password
    - username
    - uuid
    type: object
...
info:
  contact: {}
  description: |-
    基于Gin + Vue + Element UI前后端分离的工单系统的接口文档
    添加qq群: 1127401830 进入技术交流群 请备注，谢谢！
  license:
    name: MIT
    url: https://github.com/lanyulei/ferry/blob/master/LICENSE
  title: ferry API
  version: 0.0.1
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: 获取token
      parameters:
      - description: Add account
        in: body
        name: username
        required: true
        schema:
          $ref: '#/definitions/models.Login'
          type: object
      responses:
        "200":
          description: '{"code": 200, "expire": "2019-08-07T12:45:48+08:00", "token":
            ".eyJleHAiOjE1NjUxNTMxNDgsImlkIjoiYWRtaW4iLCJvcmlnX2lhdCI6MTU2NTE0OTU0OH0.-zvzHvbg0A"
            }'
          schema:
            type: string
      summary: 登陆
  /logout:
    post:
      consumes:
      - application/json
      description: 获取token
      responses:
        "200":
          description: '{"code": 200, "msg": "成功退出系统" }'
          schema:
            type: string
      security:
      - "": []
      summary: 退出登录
...
  /sd/disk:
    get:
      consumes:
      - text/html
      description: 服务器硬盘使用量 DiskCheck checks the disk usage.
      produces:
      - text/html
      responses:
        "200":
          description: 'OK - Free space: 16321MB (15GB) / 51200MB (50GB) | Used: 31%'
          schema:
            type: string
        "429":
          description: WARNING
          schema:
            type: string
        "500":
          description: CRITICAL
          schema:
            type: string
      summary: 服务器硬盘使用量
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"